requirements.txt
fastapi==0.110.1
Jinja2==3.1.3
uvicorn==0.29.0
starlette==0.37.2
Markdown==3.6
yt_iframe==1.0.5
python-multipart==0.0.9
furl==2.1.3
mybody_api_client==0.5.18
pydantic_settings==2.2.1
